+++ src/transformers/models/grounding_dino/image_processing_grounding_dino.py
@@ -111,7 +111,7 @@ def get_size_with_aspect_ratio(image_size, size, max_size=None) -> tuple[int, in
         max_original_size = float(max((height, width)))
         if max_original_size / min_original_size * size > max_size:
             raw_size = max_size * min_original_size / max_original_size
-            size = int(round(raw_size))
+            size = round(raw_size)
 
     if (height <= width and height == size) or (width <= height and width == size):
         oh, ow = height, width
