+++ tests/models/wav2vec2/test_modeling_wav2vec2.py
@@ -209,7 +209,7 @@ class Wav2Vec2ModelTester:
         output_seq_length = self.seq_length
         for kernel, stride in zip(self.conv_kernel, self.conv_stride):
             output_seq_length = (output_seq_length - (kernel - 1)) / stride
-        self.output_seq_length = int(math.ceil(output_seq_length))
+        self.output_seq_length = math.ceil(output_seq_length)
         self.encoder_seq_length = self.output_seq_length
 
         self.adapter_output_seq_length = (self.output_seq_length - 1) // adapter_stride + 1
